+++ familyhelper-clannad-impl/src/main/java/com/dwarfeng/familyhelper/clannad/impl/handler/OperateHandlerValidator.java
@@ -32,6 +32,8 @@ public class OperateHandlerValidator {
     private final CertificateMaintainService certificateMaintainService;
     private final PoceMaintainService poceMaintainService;
     private final CertificateFileInfoMaintainService certificateFileInfoMaintainService;
+    private final NotifySettingMaintainService notifySettingMaintainService;
+    private final NotifyTopicMaintainService notifyTopicMaintainService;
 
     public OperateHandlerValidator(
             UserMaintainService userMaintainService,
@@ -40,7 +42,9 @@ public class OperateHandlerValidator {
             ProfileMaintainService profileMaintainService,
             CertificateMaintainService certificateMaintainService,
             PoceMaintainService poceMaintainService,
-            CertificateFileInfoMaintainService certificateFileInfoMaintainService
+            CertificateFileInfoMaintainService certificateFileInfoMaintainService,
+            NotifySettingMaintainService notifySettingMaintainService,
+            NotifyTopicMaintainService notifyTopicMaintainService
     ) {
         this.userMaintainService = userMaintainService;
         this.avatarInfoMaintainService = avatarInfoMaintainService;
@@ -49,6 +53,8 @@ public class OperateHandlerValidator {
         this.certificateMaintainService = certificateMaintainService;
         this.poceMaintainService = poceMaintainService;
         this.certificateFileInfoMaintainService = certificateFileInfoMaintainService;
+        this.notifySettingMaintainService = notifySettingMaintainService;
+        this.notifyTopicMaintainService = notifyTopicMaintainService;
     }
 
     public void makeSureUserExists(StringIdKey userKey) throws HandlerException {
@@ -162,6 +168,26 @@ public class OperateHandlerValidator {
         try {
             if (!certificateFileInfoMaintainService.exists(certificateFileKey)) {
                 throw new CertificateFileNotExistsException(certificateFileKey);
+            }
+        } catch (ServiceException e) {
+            throw new HandlerException(e);
+        }
+    }
+
+    public void makeSureNotifySettingExists(LongIdKey notifySettingKey) throws HandlerException {
+        try {
+            if (!notifySettingMaintainService.exists(notifySettingKey)) {
+                throw new NotifySettingNotExistsException(notifySettingKey);
+            }
+        } catch (ServiceException e) {
+            throw new HandlerException(e);
+        }
+    }
+
+    public void makeSureNotifyTopicExists(StringIdKey notifyTopicKey) throws HandlerException {
+        try {
+            if (!notifyTopicMaintainService.exists(notifyTopicKey)) {
+                throw new NotifyTopicNotExistsException(notifyTopicKey);
             }
         } catch (ServiceException e) {
             throw new HandlerException(e);